Split out vector to avoid circular imports
