+++ fehm_toolkit/fehm_objects/__init__.py
@@ -1,3 +1,4 @@
 from .element import Element
 from .grid import Grid
-from .node import Node, Vector
+from .node import Node
+from .vector import Vector
