+++ fehm_toolkit/fehm_objects/node.py
@@ -1,12 +1,14 @@
 from dataclasses import dataclass
 
+from .vector import Vector
+
 
 @dataclass
 class Node:
     """Class for tracking individual node properties."""
     number: int
-    coordinates: 'Vector'
-    outside_area: 'Vector' = None
+    coordinates: Vector
+    outside_area: Vector = None
     depth: float = None
 
     @property
@@ -20,18 +22,3 @@ class Node:
     @property
     def z(self):
         return self.coordinates.z
-
-
-@dataclass
-class Vector:
-    """Class for tracking vector values (e.g. vornoi areas)."""
-    x: float
-    y: float
-    z: float
-
-    @property
-    def value(self) -> tuple[float]:
-        return (self.x, self.y, self.z)
-
-    def __format__(self, fmt) -> str:
-        return f'{self.x:{fmt}}\t{self.y:{fmt}}\t{self.z:{fmt}}'
